semaphore: Try to minimize syscalls
This commit adds an atomic variable to try to avoid syscalls on the
semaphore, so when there are "tickets" available no function with the
potential to make a syscall is called. It aims specially to avoid
unneeded calls to notify, as a non contended mutex is usally solved in
user-space too.

This check will be slightly detrimental when there are waiters, as
another CAS operation is added on top.

From the practical point of view, on this codebase it is not affecting
performance but it might help preventing spikes.

Redesigning ringbuffer seems possible but not to be a problem at
present. Maybe that queue can be traced on the future to see if it
sometimes causes spikes.